move the formatting of messages in llm_openai instead of tools. Removed llm_thread related codes from llm_tool to fully on llm_thread module
