+++ llm_openai/__init__.py
@@ -1,1 +1,2 @@
 from . import models
+from . import utils
