+++ llm_thread/views/llm_thread_views.xml
@@ -23,6 +23,11 @@
                             <field name="user_id" />
                             <field name="provider_id" />
                             <field name="model_id" />
+                            <field
+                            name="tool_ids"
+                            widget="many2many_tags"
+                            options="{'no_create': True}"
+                            />
                         </group>
                     </group>
                 </sheet>
