+++ llm_tool/__manifest__.py
@@ -30,7 +30,6 @@
     },
     "data": [
         "security/ir.model.access.csv",
-        "views/llm_thread_views.xml",
         "views/llm_tool_views.xml",
         "views/llm_tool_server_action_views.xml",
         "views/llm_tool_consent_config_views.xml",
