+++ llm_tool/models/__init__.py
@@ -1,4 +1,3 @@
-from . import llm_thread
 from . import llm_tool
 from . import llm_provider
 from . import llm_tool_record_retriever
