--- llm_tool/views/llm_thread_views.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <!-- Inherit LLM Thread Form View -->
-    <record id="view_llm_thread_form_inherit_tools" model="ir.ui.view">
-        <field name="name">llm.thread.form.inherit.tools</field>
-        <field name="model">llm.thread</field>
-        <field name="inherit_id" ref="llm_thread.llm_thread_view_form" />
-        <field name="arch" type="xml">
-            <!-- Add tools field after model_id -->
-            <xpath expr="//field[@name='model_id']" position="after">
-                <field
-          name="tool_ids"
-          widget="many2many_tags"
-          options="{'no_create': True}"
-        />
-            </xpath>
-        </field>
-    </record>
-</odoo>
